requirements.txt
twilio==7.6.0
beautifulsoup4==4.10.0
requests==2.27.1
python-dotenv==0.19.2